move run to new module preprocessing_runner.py
